pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "picoblog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bleach>=6.2.0",
    "flask>=3.1.2",
    "flask-limiter>=3.5.0",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "markdown2>=2.5.4",
    "python-dotenv>=1.0.0",
    "wtforms-sqlalchemy>=0.3",
    "email-validator>=2.0.0",
    "pygments>=2.18.0",
    "Pillow>=10.1.0",
    "gunicorn>=22.0.0"
]

[tool.setuptools]
packages = ["app"]